minor refactor for separation
